requirements.txt
faiss-cpu
torch
sentence-transformers
transformers
bitsandbytes
fastapi
streamlit
numpy
rank_bm25
accelerate>=0.26.0
yfinance
